tests: dedupe fixtures, cover grace cancel; README polish
